WebApp/requirements.txt
Flask==2.0.3
torch==1.10.0
torchvision==0.11.1
Pillow==8.4.0
numpy==1.21.2
scikit-learn==0.24.2
joblib==1.0.1
gdown==4.2.0
opencv-python==4.5.3.20210927
Flask-Cors==3.0.10
watchdog==2.1.6
jsonschema==3.2.0